==> rmds/simulations/1.0-adjust.batch_sparseDOSSA.Rmd <==
---
title: "sparseDOSSA simulation for adjust.batch function"
author: "Siyuan Ma"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
html_document:
df_print: paged
---
```{r setup, echo = FALSE}                                           
knitr::opts_knit$set(root.dir = normalizePath("../../"))
```
```{r setup2, message=FALSE, warning=FALSE, echo = FALSE}
smar::sourceDir("functions/", recursive = FALSE)
smar::sourceDir("functions/simulations/", recursive = FALSE)
setup("./")
dir_output <- "results/simulations/adjust.batch/"
dir.create(dir_output, recursive = TRUE, showWarnings = TRUE)
dir.create(paste0(dir_output, "sparseDOSSA_sets/"), recursive = TRUE, showWarnings = TRUE)
```

```{r simulation parameters}
# parallel computing
ncores <- 20
# sample size
nBatch <- c(2, 4, 6, 8)
nSample_perBatch <- c(20, 100, 500)
# feature size
nMicrobe <- c(200, 1000)
spikeMicrobes <- c(0.05, 0.1, 0.2)
# metadata specification
metadata_type <- c("batch" = "factor",
                   "positive_control1" = "factor",
                   "positive_control2" = "numeric",
                   "negative_control" = "factor")
tb_effectSize <- tidyr::crossing(batch = c(1, 2, 5, 10),
                                 positive_control1 = c(10),
                                 negative_control = 0) %>% 
  dplyr::mutate(positive_control2 = positive_control1) %>% 
  dplyr::select(batch, positive_control1, positive_control2, negative_control)
effectSize <- tibble::tibble(effectSize =
                               (1:nrow(tb_effectSize)) %>%
                               purrr::map(function(i) unlist(tb_effectSize[i, ])))
# replicates
rep <- 1:20
# metadata simulation
fDummyData <- TRUE
direction <- TRUE
# additional setup
noZeroInflate <- FALSE

tb_simSetup <- tidyr::crossing(
  nBatch,
  nSample_perBatch,
  nMicrobe,
  spikeMicrobes,
  noZeroInflate,
  effectSize
) %>% 
  dplyr::mutate(i_setup = 1:dplyr::n()) %>% 
  tidyr::crossing(rep) %>% 
  dplyr::mutate(i = 1:dplyr::n(),
                nSample = nSample_perBatch * nBatch)
```

```{r simulate metadata}
set.seed(1)
tb_sim <- tb_simSetup %>% 
  dplyr::group_by(i) %>% 
  dplyr::mutate(
    df_metadata = 
      data.frame(batch = create_batch(nSample_perBatch = nSample_perBatch,
                                      nBatch = nBatch),
                 positive_control1 = rbinom(n = nSample, size = 1, prob = 0.5) %>% 
                   as.factor(),
                 positive_control2 = runif(n = nSample, min = -1, max = 1),
                 negative_control = rbinom(n = nSample, size = 1, prob = 0.5) %>% 
                   as.factor(),
                 row.names = paste0("Sample", 1:nSample)) %>% 
      list(),
    spikein.mt = 
      create_effectSize(effectSize = effectSize[[1]],
                        df_metadata = df_metadata[[1]],
                        metadata_type = metadata_type,
                        fDummyData = fDummyData) %>%
      create_spikein.mt(number_features = nMicrobe,
                        percent_spiked = spikeMicrobes,
                        effectSize = .,
                        direction = direction,
                        seed = rep) %>%
      list()) %>% 
  dplyr::ungroup()
save(tb_sim, file = paste0(dir_output, "tb_sim.RData"))
```

```{r run sparseDOSSA}
set.seed(1)
N <- nrow(tb_sim)
start.time <- Sys.time()
doParallel::registerDoParallel(ncores)
l_sparseDOSSA <- foreach::`%dopar%`(
  foreach::foreach(i = 1:N),
  {
    i_simSetup <- tb_sim[i, ]
    UserMetadata <- create_metadataMatrix(df_metadata = i_simSetup$df_metadata[[1]],
                                          metadata_type = metadata_type,
                                          fDummyData = fDummyData,
                                          scale = FALSE)
    if(nrow(UserMetadata) != max(i_simSetup$spikein.mt[[1]]$metadata))
      stop("Number of metadata variables in UserMetadata does not agree with spikein matrix!",
           " Something must've gone wrong!")
    
    otu_count <- sparseDOSSA::sparseDOSSA(
      number_features = i_simSetup$nMicrobe,
      number_samples = i_simSetup$nSample,
      UserMetadata = UserMetadata,
      spikein.mt = format_spikein.mt(i_simSetup$spikein.mt[[1]]),
      noZeroInflate = i_simSetup$noZeroInflate,
      percent_spiked = i_simSetup$spikeMicrobes,
      seed = i_simSetup$i,
      minLevelPercent = 0,
      write_table = FALSE,
      verbose = FALSE) %>%
      extract_sparseDOSSA() %>%
      extract2("features")
    
    otu_count %>% data.frame() %>% 
      cbind(data.frame(Features = rownames(otu_count)), .) %>% 
      readr::write_tsv(path = paste0(dir_output, "sparseDOSSA_sets/", i, ".tsv"))
    return(NULL)
  }
)
doParallel::stopImplicitCluster()
# save(l_sparseDOSSA, file = paste0(dir_output, "sparseDOSSA.RData"))
print(Sys.time() - start.time)
```
