==> rmds/simulations/1.1-adjust.batch_evaluate.Rmd <==
---
title: "Evaluation for adjust.batch function"
author: "Siyuan Ma"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
html_document:
df_print: paged
---
```{r setup, echo = FALSE}                                           
knitr::opts_knit$set(root.dir = normalizePath("../../"))
```
```{r setup2, message=FALSE, warning=FALSE, echo = FALSE}
smar::sourceDir("functions/", recursive = FALSE)
smar::sourceDir("functions/simulations/", recursive = FALSE)
setup("./")
dir_output <- "results/simulations/adjust.batch/"
```

```{r load data}
# parallel computing
ncores <- 20
load(paste0(dir_output, "tb_sim.RData"))
```

```{r calculate R2 before adjustment}
N <- nrow(tb_sim)
start.time <- Sys.time()
doParallel::registerDoParallel(ncores)
results <-  foreach::`%dopar%` (
  foreach::foreach(i = 1:N), 
  {
    cat(i, "\n", file = paste0(dir_output, "evaluate_before_progress.txt"), append = TRUE)
    mat_otu <- read.table(paste0(dir_output,
                                 "sparseDOSSA_sets/",
                                 i, ".tsv"),
                          header = TRUE,
                          sep = "\t",
                          row.names = 1) %>%
      as.matrix
    i_simSetup <- tb_sim[i, ]
    df_metadata <- i_simSetup$df_metadata[[1]]
    
    phylo <- phyloseq::phyloseq(
      phyloseq::otu_table(mat_otu, taxa_are_rows = TRUE),
      phyloseq::sample_data(df_metadata)) %>%
      phyloseq::transform_sample_counts(MMUPHin:::tss) %>%
      smar::prune_taxaSamples()
    D <- phyloseq::distance(phylo, method = "bray")
    R2_batch <- vegan::adonis(D ~ batch,
                              data = smar::sample_data2(phylo),
                              permutations = 2)$aov.tab["batch", "R2"]
    R2_positive_control1 <- vegan::adonis(D ~ positive_control1,
                                          data = smar::sample_data2(phylo),
                                          permutations = 2)$aov.tab["positive_control1", "R2"]
    R2_positive_control2 <- vegan::adonis(D ~ positive_control2,
                                          data = smar::sample_data2(phylo),
                                          permutations = 2)$aov.tab["positive_control2", "R2"]
    R2_negative_control <- vegan::adonis(D ~ negative_control,
                                         data = smar::sample_data2(phylo),
                                         permutations = 2)$aov.tab["negative_control", "R2"]
    
    
    return(list(R2 = c("batch" = R2_batch,
                       "positive_control1" = R2_positive_control1,
                       "positive_control2" = R2_positive_control2,
                       "negative_control" = R2_negative_control)))
  }
)
doParallel::stopImplicitCluster()
save(results, file = paste0(dir_output, "evaluate_before.RData"))
print(Sys.time() - start.time)
```

```{r batch adjustment}
start.time <- Sys.time()
doParallel::registerDoParallel(ncores)
results_correct <- foreach::`%dopar%`(
  foreach::foreach(i = 1:N),
  {
    cat(i, "\n", file = paste0(dir_output, "evaluate_after_progress.txt"), append = TRUE)
    mat_otu <- read.table(paste0(dir_output,
                                 "sparseDOSSA_sets/",
                                 i, ".tsv"),
                          header = TRUE,
                          sep = "\t",
                          row.names = 1) %>%
      as.matrix
    i_simSetup <- tb_sim[i, ]
    df_metadata <- i_simSetup$df_metadata[[1]]
    
    mat_otu <- mat_otu %>%
      MMUPHin::adjust.batch(batch = "batch",
                            covariates = c("positive_control1",
                                           "positive_control2",
                                           "negative_control"),
                            data = df_metadata,
                            diagnostics = FALSE,
                            verbose = FALSE)
    
    phylo <- phyloseq::phyloseq(
      phyloseq::otu_table(mat_otu, taxa_are_rows = TRUE),
      phyloseq::sample_data(df_metadata)) %>%
      phyloseq::transform_sample_counts(MMUPHin:::tss) %>%
      smar::prune_taxaSamples()
    D <- phyloseq::distance(phylo, method = "bray")
    R2_batch <- vegan::adonis(D ~ batch,
                              data = smar::sample_data2(phylo),
                              permutations = 2)$aov.tab["batch", "R2"]
    R2_positive_control1 <- vegan::adonis(D ~ positive_control1,
                                          data = smar::sample_data2(phylo),
                                          permutations = 2)$aov.tab["positive_control1", "R2"]
    R2_positive_control2 <- vegan::adonis(D ~ positive_control2,
                                          data = smar::sample_data2(phylo),
                                          permutations = 2)$aov.tab["positive_control2", "R2"]
    R2_negative_control <- vegan::adonis(D ~ negative_control,
                                         data = smar::sample_data2(phylo),
                                         permutations = 2)$aov.tab["negative_control", "R2"]
    
    
    return(list(R2 = c("batch" = R2_batch,
                       "positive_control1" = R2_positive_control1,
                       "positive_control2" = R2_positive_control2,
                       "negative_control" = R2_negative_control))
    )
  }
) 
doParallel::stopImplicitCluster()
save(results_correct, file = paste0(dir_output, "evaluate_after.RData"))
print(Sys.time() - start.time)
```

```{r ComBat}
start.time <- Sys.time()
doParallel::registerDoParallel(ncores)
results_ComBat <- foreach::`%dopar%`(
  foreach::foreach(i = 1:N),
  {
    cat(i, "\n", file = paste0(dir_output, "evaluate_ComBat_progress.txt"), append = TRUE)
    mat_otu <- read.table(paste0(dir_output,
                                 "sparseDOSSA_sets/",
                                 i, ".tsv"),
                          header = TRUE,
                          sep = "\t",
                          row.names = 1) %>%
      as.matrix
    i_simSetup <- tb_sim[i, ]
    df_metadata <- i_simSetup$df_metadata[[1]]
    
    mat_otu <- mat_otu %>%
      MMUPHin::adjust.batch(
        batch = "batch",
        covariates = c("positive_control1",
                       "positive_control2",
                       "negative_control"),
        data = df_metadata,
        zero.inflation = FALSE,
        diagnostics = FALSE,
        verbose = FALSE
      )
    
    phylo <- phyloseq::phyloseq(
      phyloseq::otu_table(mat_otu, taxa_are_rows = TRUE),
      phyloseq::sample_data(df_metadata)) %>%
      phyloseq::transform_sample_counts(MMUPHin:::tss) %>%
      smar::prune_taxaSamples()
    D <- phyloseq::distance(phylo, method = "bray")

    R2_batch <- vegan::adonis(D ~ batch,
                              data = smar::sample_data2(phylo),
                              permutations = 2)$aov.tab["batch", "R2"]
    R2_positive_control1 <- vegan::adonis(D ~ positive_control1,
                                          data = smar::sample_data2(phylo),
                                          permutations = 2)$aov.tab["positive_control1", "R2"]
    R2_positive_control2 <- vegan::adonis(D ~ positive_control2,
                                          data = smar::sample_data2(phylo),
                                          permutations = 2)$aov.tab["positive_control2", "R2"]
    R2_negative_control <- vegan::adonis(D ~ negative_control,
                                         data = smar::sample_data2(phylo),
                                         permutations = 2)$aov.tab["negative_control", "R2"]
    
    
    return(list(R2 = c("batch" = R2_batch,
                       "positive_control1" = R2_positive_control1,
                       "positive_control2" = R2_positive_control2,
                       "negative_control" = R2_negative_control))
    )
  }
)  
doParallel::stopImplicitCluster()
save(results_ComBat, file = paste0(dir_output, "evaluate_ComBat.RData"))
print(Sys.time() - start.time)
```
